pyproject.toml
[project]
name = "openinsider-notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.3",
    "matrix-nio>=0.25.1",
]
